+++ test/run-tests.js
@@ -2,3 +2,4 @@
 
 require("./hostlink-core.test");
 require("./hostlink-high-level.test");
+require("./node-runtime.test");